Refactoring API calls for audioscrobbler compatible services
Refactoring build request functions to use the full api_credentials struct

Minor cleanups
