+++ src/signon.c
@@ -65,9 +65,9 @@ static void get_session(struct api_credentials *creds)
 
     CURL *curl = curl_easy_init();
     struct http_response *res = http_response_new();
-    struct http_request *req = api_build_request_get_session(curl, creds->end_point, creds->token);
+    struct http_request *req = api_build_request_get_session(curl, creds);
 
-    enum api_return_statuses ok = api_get_request(curl, req, res);
+    enum api_return_status ok = api_get_request(curl, req, res);
     curl_easy_cleanup(curl);
     http_request_free(req);
 
@@ -91,9 +91,9 @@ static void get_token(struct api_credentials *creds)
 
     CURL *curl = curl_easy_init();
     struct http_response *res = http_response_new();
-    struct http_request *req = api_build_request_get_token(curl, creds->end_point);
+    struct http_request *req = api_build_request_get_token(curl, creds);
 
-    enum api_return_statuses ok = api_get_request(curl, req, res);
+    enum api_return_status ok = api_get_request(curl, req, res);
     curl_easy_cleanup(curl);
     http_request_free(req);
